requirements.txt
absl-py==1.2.0
acme==1.1.0
albumentations==1.3.0
anyio==3.6.2
apturl==0.5.2
asgiref==3.5.2
astroid==2.12.5
async-generator==1.10
attrs==21.2.0
autopep8==1.4.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
certbot==0.40.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.8
chumpy==0.70
Click==7.0
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
contourpy==1.0.5
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.32
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
dill==0.3.5.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.8
django-allauth==0.39.1
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11
django-countries==5.3.3
django-crispy-forms==1.7.2
django-datetimepicker==3.14
django-debug-toolbar==1.10.1
django-filter==22.1
django-mongoengine==0.5.4
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djangotoolbox==1.8.0
dnspython==2.2.1
easydict==1.9
entrypoints==0.3
evdev==1.6.0
exceptiongroup==1.0.1
execnet==1.9.0
face-alignment==1.3.5
fastapi==0.85.1
flatbuffers==2.0.7
fonttools==4.37.3
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.11.1
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.48.1
h11==0.14.0
httplib2==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.3
imageio==2.4.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
iniconfig==1.1.1
insightface==0.2.1
install==1.3.5
iopath==0.1.10
isort==5.10.1
joblib==1.2.0
josepy==1.2.0
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.4.4
kornia==0.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
llvmlite==0.39.1
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
mccabe==0.7.0
mock==3.0.5
mongoengine==0.24.2
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.2.1
names==0.3.0
netifaces==0.10.4
networkx==2.8.6
ninja==1.10.2.4
numba==0.56.2
numpy==1.23.3
oauthlib==3.0.1
olefile==0.46
onnx==1.12.0
onnxruntime==1.12.1
onnxruntime-gpu==1.12.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.5.0
parsedatetime==2.4
pbr==5.4.5
pep8==1.7.1
pexpect==4.6.0
Pillow==8.1.2
platformdirs==2.5.2
playwright==1.19.1
pluggy==1.0.0
portalocker==2.5.1
prettytable==3.4.1
proglog==0.1.10
protobuf==3.19.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
pydantic==1.10.2
pyee==8.1.0
PyGetWindow==0.0.9
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pylint==2.15.0
pymacaroons==0.13.0
pymongo==4.3.2
PyMsgBox==1.0.9
PyNaCl==1.3.0
pynput==1.7.6
pyOpenSSL==19.0.0
pyparsing==3.0.6
pyperclip==1.8.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.0.1
pytest-base-url==1.4.2
pytest-forked==1.4.0
pytest-playwright==0.2.3
pytest-xdist==2.5.0
python-apt==2.0.0+ubuntu0.20.4.8
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.1
python-dotenv==0.21.0
python-multipart==0.0.5
python-slugify==5.0.2
python-xlib==0.32
python3-openid==3.1.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.2.1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==5.1.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
qudida==0.0.4
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.2.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.5.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
ssh-import-id==5.10
starlette==0.20.4
stripe==2.27.0
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
torch==1.6.0
torch-tb-profiler==0.4.0
torchaudio==0.12.1
torchvision==0.7.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.3.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
uvicorn==0.19.0
uvloop==0.17.0
vboxapi==1.0
wadllib==1.3.3
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xkit==0.0.0
yacs==0.1.8
zipp==3.8.1
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
